refactor(canvas): #1032 CanvasLayout.tsx を hook/コンポーネントへ分割 (god-file 解消)
- チーム起動 (builtin/保存済み/custom leader/recent 復元 + 音声指揮ラッパ) を lib/hooks/use-canvas-spawn.ts へ切り出し (spawnTeam 経路統一の不変条件を単一モジュールに閉じる)
- 単体カード追加 (agent/custom/API/editor 等 + stagger 配置規則) を lib/hooks/use-canvas-add-card.ts へ切り出し
- AppMenuBar/Topbar 操作 (workspace 系 / 再起動 / 更新 / Clear) を lib/hooks/use-canvas-menu-actions.ts へ切り出し
- チーム起動 FAB + popover (開閉/タブ/保存済みプリセット取得) を components/canvas/CanvasSpawnFab.tsx へ切り出し
- 死蔵状態だった addCardOpen/addPopoverRef と未使用 import (AddItem/AgentBadge/icons) を削除
- CanvasLayout.tsx: 1009 -> 257 行。baseline から CanvasLayout.tsx / i18n.ts を除去 (--update-baseline)
